pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evc"
version = "0.1.0"
authors = [
  { name="Filipe Verri", email="filipeverri@gmail.com" },
]
description = "Tools to evaluate new versions of pipelines to generate machine learning models."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "numpy >= 1.21",
  "scipy",
  "matplotlib",
  "cmcrameri",
  "pandas",
  "GitPython"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/verri/evc"
Issues = "https://github.com/verri/evc/issues"
